+++ app/forgot-password/page.tsx
@@ -1,0 +1,5 @@
+import ScreenForgotPassword from '@/components/screen/ForgotPassword';
+
+export default function ForgotPassword() {
+	return <ScreenForgotPassword />;
+}
